u_i should be type int not short - should be 4 bytes

+++ cbflib_adaptbx/detectors/buffer_based_service.cpp
@@ -12,7 +12,7 @@ typedef union {
 
 typedef union {
   char b[4];
-  short i;
+  int i;
 } u_i;
 
 // functions for byte swapping
